First round of refactoring and fixing
