+++ omnimatter_marathon/locale/en/en.cfg
@@ -1,12 +1,16 @@
 [mod-setting-name]
 
 omnimarathon_exponential=Exponential omnimarathon increase.
-omnimarathon_constant=Constant omnimarathon increase.
+omnimarathon_constant=Linear / Exponential increase.
 omnimarathon_time_increase=Omnimarathon time increase.
 omnimarathon_time_const=Constant time increase in omnimarathon.
+omnimarathon_adapt_fuel_value=Adapt Fuel value
+omnimarathon_rounding=Rounding constant?
 
 [mod-setting-description]
 omnimarathon_exponential=Decides whether the increase in cost is exponential or linear. Use linear for a somewhat easier omnimarathon experience.
-omnimarathon_constant=The increase in cost for each individual item. This cost is applied in both linear and exponential omnimarathon modes. If this value is above 2, issues are guaranteed to happen in exponential mode.
+omnimarathon_constant=The increase in cost for each individual item. This cost is applied in both linear and exponential modes. If this value is above 2, issues are guaranteed to happen in exponential mode.
 omnimarathon_time_increase=Decides whether crafting speed is increased or not.
 omnimarathon_time_const=Determines the increase in time. This increase is percentual (0.1 = 10% increase)
+omnimarathon_adapt_fuel_value=Adapt Fuel value
+omnimarathon_rounding=Rounding constant?
