+++ omnimatter_marathon/settings.lua
@@ -4,7 +4,7 @@ data:extend(
     type = "bool-setting",
     name = "omnimarathon_exponential",
     setting_type = "startup",
-    default_value = false,
+    default_value = true,
 	order=a
   },
   {
@@ -12,35 +12,35 @@ data:extend(
     name = "omnimarathon_constant",
     setting_type = "startup",
     default_value = "1/2",
-	order=a
+	order=b
   },
   {
     type = "bool-setting",
     name = "omnimarathon_time_increase",
     setting_type = "startup",
     default_value = true,
-	order=a
-  },
-  {
-    type = "bool-setting",
-    name = "omnimarathon_adapt_fuel_value",
-    setting_type = "startup",
-    default_value = true,
-	order=a
+	order=c
   },
   {
     type = "double-setting",
     name = "omnimarathon_time_const",
     setting_type = "startup",
     default_value = 0,
-	order=a
+	order=e
+  },
+  {
+    type = "bool-setting",
+    name = "omnimarathon_adapt_fuel_value",
+    setting_type = "startup",
+    default_value = true,
+	order=d
   },
   {
     type = "bool-setting",
     name = "omnimarathon_rounding",
     setting_type = "startup",
     default_value = false,
-	order=a
+	order=f
   },
 }
 )
